Move common portions to python AI Timer into common.
This refactors AI/freeorion_tools/timers and moves the portions that do
not depend in freeOrionAIInterface into python/common.

It renames the FO AI dependent timer to FOTimer.
