+++ default/python/AI/ColonisationAI.py
@@ -12,10 +12,10 @@ import TechsListsAI
 import MilitaryAI
 from turn_state import state
 from EnumsAI import MissionType, FocusType, EmpireProductionTypes, ShipRoleType, PriorityType
-from freeorion_tools import dict_from_map, tech_is_complete, get_ai_tag_grade, cache_by_turn, Timer
+from freeorion_tools import dict_from_map, tech_is_complete, get_ai_tag_grade, cache_by_turn, FOTimer
 from AIDependencies import INVALID_ID, POP_CONST_MOD_MAP, POP_SIZE_MOD_MAP
 
-colonization_timer = Timer('getColonyFleets()')
+colonization_timer = FOTimer('getColonyFleets()')
 
 
 empire_colonizers = {}
