+++ default/python/AI/FreeOrionAI.py
@@ -28,13 +28,13 @@ import ResearchAI
 import ResourcesAI
 import TechsListsAI
 from AIDependencies import INVALID_ID
-from freeorion_tools import UserStringList, chat_on_error, print_error, UserString, handle_debug_chat, Timer, init_handlers
+from freeorion_tools import UserStringList, chat_on_error, print_error, UserString, handle_debug_chat, FOTimer, init_handlers
 from common.listeners import listener
 from character.character_module import Aggression
 from character.character_strings_module import get_trait_name_aggression, possible_capitals
 
-main_timer = Timer('timer', write_log=True)
-turn_timer = Timer('bucket', write_log=True)
+main_timer = FOTimer('timer', write_log=True)
+turn_timer = FOTimer('bucket', write_log=True)
 
 using_statprof = False
 try:
