+++ default/python/AI/freeorion_tools/__init__.py
@@ -1,5 +1,5 @@
 from _freeorion_tools import *
 from extend_freeorion_AI_interface import patch_interface
 from interactive_shell import handle_debug_chat
-from timers import Timer
+from timers import FOTimer
 from handlers import init_handlers
