+++ default/python/AI/freeorion_tools/timers.py
@@ -1,6 +1,6 @@
 import os
 import freeOrionAIInterface as fo
-from time import time
+from common.timers import Timer
 from common.option_tools import get_option_dict, check_bool, DEFAULT_SUB_DIR
 from common.option_tools import TIMERS_TO_FILE, TIMERS_USE_TIMERS, TIMERS_DUMP_FOLDER
 import sys
@@ -30,7 +30,7 @@ def _get_timers_dir():
 
 class DummyTimer(object):
     """
-    Dummy timer to be called if no need any logging.
+    Dummy timer to be used if timers are disabled.
     """
     def __init__(self, *args, **kwargs):
         pass
@@ -45,39 +45,18 @@ class DummyTimer(object):
         pass
 
 
-class LogTimer(object):
+class FOLogTimer(Timer):
+    """A Timer with a FO AI engine dependent extension that logs timer results to a file each turn.
+    """
     def __init__(self, timer_name, write_log=False):
         """
         Creates timer. Timer name is name that will be in logs header and part of filename if write_log=True
         If write_log true and timers logging enabled (DUMP_TO_FILE=True) save timer info to file.
 
         """
-        self.timer_name = timer_name
-        self.start_time = None
-        self.section_name = None
-        self.timers = []
-        self.write_log = write_log
+        Timer.__init__(self, timer_name)
         self.headers = None
-        self.log_name = None
-
-    def stop(self, section_name=''):
-        """
-        Stop timer if running. Specify section_name if want to override its name.
-        """
-        if self.start_time:
-            self.end_time = time()
-            section_name = section_name or self.section_name
-            self.timers.append((section_name, (self.end_time - self.start_time) * 1000.0))
-        self.start_time = None
-        self.section_name = None
-
-    def start(self, section_name):
-        """
-        Stop prev timer if present and starts new.
-        """
-        self.stop()
-        self.section_name = section_name
-        self.start_time = time()
+        self.write_log = write_log
 
     def _write(self, text):
         if not _get_timers_dir():
@@ -97,21 +76,10 @@ class LogTimer(object):
         Stop timer, output result, clear checks.
         If dumping to file, if headers are not match to prev, new header line will be added.
         """
-        turn = fo.currentTurn()
-        self.stop()
-        if not self.timers:
-            return
-        max_header = max(len(x[0]) for x in self.timers)
-        line_max_size = max_header + 14
-        print
-        print ('Timing for %s:' % self.timer_name)
-        print '=' * line_max_size
-        for name, val in self.timers:
-            print "%-*s %8d msec" % (max_header, name, val)
-        print '-' * line_max_size
-        print ("Total: %8d msec" % sum(x[1] for x in self.timers)).rjust(line_max_size)
+        Timer.end(self)
 
         if self.write_log and DUMP_TO_FILE:
+            turn = fo.currentTurn()
             headers = make_header('Turn', *[x[0] for x in self.timers])
             if self.headers != headers:
                 self._write(''.join(headers) + '\n' + ''.join(['-' * (len(x) - 2) + '  ' for x in headers]))
@@ -123,4 +91,4 @@ class LogTimer(object):
             self._write(''.join(row))
         self.timers = []  # clear times
 
-Timer = LogTimer if USE_TIMERS else DummyTimer
+FOTimer = FOLogTimer if USE_TIMERS else DummyTimer
